… Ryujinx.Tests/Cpu/CpuTestSimd.cs
… Ryujinx.Tests/Cpu/CpuTestSimdArithmetic.cs