Tests confirm the functioning of the implementations (at basic level of FP Emu.), also in the handling of the zeroes, the infinites and the QNaNs. SNaNs are not converted to QNaNs, in line with other non-SoftFloat implementations.

Tests:
```
F_Rint_AMNPZ: Total tests: 12200. Passed: 12200. Failed: 0. Skipped: 0.
Test execution was successful.
Test execution time: 3,1808 Minutes

F_Rint_IX: Total tests: 19520. Passed: 19520. Failed: 0. Skipped: 0.
Test execution was successful.
Test execution time: 5,7869 Minutes
```

Performance enhancements:
```
F_Rint_MNPZ_S (Sse.Off -> Sse.On): 2908 ms -> 2096 ms (cumulative times on 2176 tests)
F_Rint_MNPZ_V (Sse.Off -> Sse.On): 8776 ms -> 5567 ms (cumulative times on 6144 tests)

F_Rint_IX_S_S (Sse.Off -> Sse.On):  6649 ms ->  4896 ms (cumulative times on  4352 tests)
F_Rint_IX_S_V (Sse.Off -> Sse.On): 21042 ms -> 14131 ms (cumulative times on 12288 tests)
```

Code JITed for some instructions:
`FRINTZ S0, S1`: https://pastebin.com/wFutCPgN
`FRINTZ V0.2S, V0.2S`: https://pastebin.com/ADu2PQ50
`FRINTI S0, S1`: https://pastebin.com/vBe20h8W
`FRINTI V0.2S, V0.2S`: https://pastebin.com/n7thtMy9

(.NET Core SDK 2.1.506)

Closes #251